+++ src/telemetry/index.ts
@@ -1,4 +1,6 @@
 export { TelemetryProvider, useTelemetry } from "./TelemetryProvider";
 export type { TelemetryProviderProps } from "./TelemetryProvider";
 export { consoleTelemetryAdapter } from "./console-adapter";
+export { createSentryTelemetryAdapter } from "./sentry-adapter";
+export type { CreateSentryTelemetryAdapterOptions, SentryLike } from "./sentry-adapter";
 export type { TelemetryAdapter, TelemetryEvent, TelemetryUser } from "./types";